refactor: 폴더, 파일, 변수 이름 수정
