--- src/main/java/clofi/codeython/evaluate/EvaluateRequest.java
@@ -1,9 +1,0 @@
-package clofi.codeython.evaluate;
-
-import lombok.Getter;
-
-@Getter
-public class EvaluateRequest {
-    private Language language;
-    private String code;
-}
