+++ src/main/java/clofi/codeython/judge/dto/JudgeRequest.java
@@ -1,0 +1,10 @@
+package clofi.codeython.judge.dto;
+
+import clofi.codeython.judge.domain.Language;
+import lombok.Getter;
+
+@Getter
+public class JudgeRequest {
+    private Language language;
+    private String code;
+}
